+++ example/lib/views/events_planner_rtl_example.dart
@@ -1,4 +1,4 @@
-import 'package:example/main.dart';
+import 'package:example/app.dart';
 import 'package:flutter/material.dart';
 import 'package:infinite_calendar_view/infinite_calendar_view.dart';
 import 'package:intl/intl.dart' as intl;